Merge phase2 updates (#621)
* integrate with the latest phase2 classifier

the latest phase2 classifier takes the test area image as the input and output the activation scores for the 5 output classes (both, flu-a, flu-b, invalid and negative). The prediction is the class with the highest score.

To make the demo available asap, this commit contains the minimal changes necessary to integrate the new classification model with the app. Further refactoring is necessary from a code perspective.

* fix RDT interpretation message

delete the old tflite file

* prevent the previously captured strip result affecting the next captured strip
fix a typo

* fix a negative byte value issue during converting bitmap to float image tensor

also optimize the pixel index update by using only incrementally addition instead of multiplication and addition, which saves 2x300x300 multiplications per inference where 300x300 is the image size.

* remove an unneeded target surface for still image capture

the surface causes an exception of unconfigured output surface on Samsung devices with Android 8.

Co-authored-by: Yongshao Ruan <yongshao@yongshaos-mbp.lan>